Refactor: "Refactor project"
Move files and folder
Change code styles
- set type define top to bottom
