+++ src/App.tsx
@@ -1,21 +1,21 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import LandingPage from './views/LandingPage';
-import LoginPage from './views/LoginPage';
-import RegisterPage from './views/RegisterPage';
-import { Notice, Chart, Benefit } from './views/company';
-import MainPage from './views/MainPage';
-import Attendance from './views/AttendancePage';
-import ProjectList from './views/ProjectList';
-import GalleryPage from './views/GalleryPage';
-import NoticeWritePage from './views/company/notice/NoticeWritePage';
-import Contents from './views/Contents';
-import NoticeUpdatePage from './views/company/notice/NoticeUpdatePage';
+import LandingPage from './pages/LandingPage';
+import LoginPage from './pages/LoginPage';
+import RegisterPage from './pages/RegisterPage';
+import { Notice, Chart, Benefit } from './pages/company';
+import MainPage from './pages/MainPage';
+import Attendance from './pages/AttendancePage';
+import ProjectList from './pages/ProjectList';
+import GalleryPage from './pages/GalleryPage';
+import NoticeWritePage from './pages/company/notice/NoticeWritePage';
+import Contents from './pages/Contents';
+import NoticeUpdatePage from './pages/company/notice/NoticeUpdatePage';
 
 import { useDispatch } from 'react-redux';
 import { login } from './store/loginSlice';
-import ProjectWirte from './views/ProjectWirte';
-import ProjectDetail from './views/ProjectDetail';
-import ProjectUpdate from './views/ProjectUpdate';
+import ProjectWirte from './pages/ProjectWirte';
+import ProjectDetail from './pages/ProjectDetail';
+import ProjectUpdate from './pages/ProjectUpdate';
 
 const App = () => {
   const isLogin = localStorage.getItem('isLogin');
