+++ src/components/Auth/SubmitButton.tsx
@@ -1,10 +1,5 @@
 import '../../scss/components/auth/submitButton.scss';
 
-interface SubmitButtonProps {
-  content: string;
-  activate: boolean;
-}
-
 const SubmitButton = ({ content, activate }: SubmitButtonProps): JSX.Element => {
   return (
     <button type="submit" className="authForm__button" disabled={activate ? false : true}>
@@ -14,3 +9,8 @@ const SubmitButton = ({ content, activate }: SubmitButtonProps): JSX.Element => 
 };
 
 export default SubmitButton;
+
+interface SubmitButtonProps {
+  content: string;
+  activate: boolean;
+}
