+++ src/components/Placeholder.tsx
@@ -1,11 +1,3 @@
-import React from 'react';
-
-interface PlaceholderProps {
-  type?: 'attendance' | '';
-  // type?: 'attendance' | '';
-  // 사용하는 곳에서 특정 스타일을 추가하고싶을 경우 type으로 넘겨서 props값 추가해주기
-}
-
 const Placeholder = ({ type }: PlaceholderProps): JSX.Element => {
   return (
     <div className={`placeholder ${type ? type : ''}`}>
@@ -15,3 +7,9 @@ const Placeholder = ({ type }: PlaceholderProps): JSX.Element => {
 };
 
 export default Placeholder;
+
+interface PlaceholderProps {
+  type?: 'attendance' | '';
+  // type?: 'attendance' | '';
+  // 사용하는 곳에서 특정 스타일을 추가하고싶을 경우 type으로 넘겨서 props값 추가해주기
+}
