--- src/firebase/fireStore.ts
@@ -1,1 +1,0 @@
-import app from './config';
