+++ src/firebase/index.ts
@@ -1,4 +1,3 @@
-// import React from 'react';
 import app from './config';
 import { getFirestore, collection } from 'firebase/firestore';
 import { getStorage, ref } from 'firebase/storage';
