+++ src/pages/LandingPage.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Outlet } from 'react-router-dom';
 
 import Header from '../components/Header';
