+++ src/pages/MainGalleryContent.tsx
