+++ src/pages/ProjectDetail.tsx
