+++ src/pages/company/notice/Notice.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import NoticeList from '../../../components/Notice/NoticeList';
 
 const Notice = () => {
