+++ src/pages/company/notice/NoticeWritePage.tsx
@@ -1,44 +1,60 @@
-import React, { useState, useEffect } from 'react';
+import { useState } from 'react';
 import { storage, db } from '../../../firebase';
-import { doc, updateDoc, getDoc } from 'firebase/firestore';
+import { doc, addDoc, collection, updateDoc } from 'firebase/firestore';
 import { ref, uploadBytes, getDownloadURL } from 'firebase/storage';
-import { useLocation } from 'react-router-dom';
+import { useSelector } from 'react-redux';
 
 import '../../../scss/components/writePage/writePage.scss';
 
-const NoticeUpdatePage = () => {
-  const location = useLocation();
-  const itemId = location.state;
+const NoticeWritePage = () => {
   const [title, setTitle] = useState<string>('');
   const [file, setFile] = useState<File | null>(null);
   const [content, setContent] = useState<string>();
+  const userEmail = useSelector(state => state.loginUpdate.email);
   const [btnDisabled, setBtnDisabled] = useState(false);
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
+
     try {
       setBtnDisabled(true);
       if (file !== null) {
         const noticeRef = ref(storage, `images/notice/${new Date().getTime() + file.name}`);
         const snapshot = await uploadBytes(noticeRef, file);
         const imgUrl = await getDownloadURL(snapshot.ref);
-        const docRef = doc(db, 'notice', itemId);
-        await updateDoc(docRef, {
+        const data = {
           title,
           content,
           time: new Date().toLocaleString(),
+          userEmail,
           url: imgUrl,
-        });
-        alert('완료 되었습니다.');
-      }
+          number: new Date().getTime(),
+        };
 
-      const docRef = doc(db, 'notice', itemId);
-      await updateDoc(docRef, {
-        title,
-        content,
-        time: new Date().toLocaleString(),
-      });
-      alert('완료 되었습니다.');
+        await addDoc(collection(db, 'notice'), data).then(item => {
+          const docRef = doc(db, 'notice', item.id);
+          updateDoc(docRef, {
+            id: item.id,
+          });
+          alert('사진 등록이 완료 되었습니다.');
+        });
+      } else {
+        const data = {
+          title,
+          content,
+          time: new Date().toLocaleString(),
+          userEmail,
+          url: null,
+          number: new Date().getTime(),
+        };
+        await addDoc(collection(db, 'notice'), data).then(item => {
+          const docRef = doc(db, 'notice', item.id);
+          updateDoc(docRef, {
+            id: item.id,
+          });
+          alert('완료 되었습니다.');
+        });
+      }
     } catch {
       setBtnDisabled(false);
       console.error();
@@ -60,23 +76,6 @@ const NoticeUpdatePage = () => {
     setFile(e.target.files && e.target.files[0]);
   };
 
-  const getData = async () => {
-    const docRef = doc(db, 'notice', itemId);
-    const docSnap = await getDoc(docRef);
-
-    if (docSnap.exists()) {
-      const title = docSnap.data().title;
-      const content = docSnap.data().content;
-      setTitle(title);
-      setContent(content);
-    } else {
-      console.error('파일이 없습니다!');
-    }
-  };
-  useEffect(() => {
-    getData();
-  }, []);
-
   return (
     <form className="write-form" onSubmit={handleSubmit}>
       <div className="write-form__title">
@@ -86,16 +85,11 @@ const NoticeUpdatePage = () => {
           className="write-form__title__item"
           placeholder="제목을 입력하세요"
           onChange={handleTitle}
-          value={title}
           required></input>
         ;
       </div>
       <div className="write-form__content">
-        <textarea
-          name="content"
-          className="write-form__content__item"
-          onChange={handleContent}
-          value={content}></textarea>
+        <textarea name="content" className="write-form__content__item" onChange={handleContent}></textarea>
       </div>
       <div className="write-util">
         <input type="file" accept="image/*" name="file" className="write-util__image" onChange={handleFile} />
@@ -107,4 +101,4 @@ const NoticeUpdatePage = () => {
   );
 };
 
-export default NoticeUpdatePage;
+export default NoticeWritePage;
