+++ src/scss/components/_headerDropDown.scss
@@ -17,6 +17,7 @@
     & li {
       padding: 0.5rem;
       cursor: pointer;
+      font-size: 1.2rem;
       &a {
         display: inline-block;
       }
