+++ src/scss/components/_sidebar.scss
@@ -3,7 +3,7 @@
 
 .nav {
   width: 224px;
-  height: calc(100vh - 60px);
+  height: auto;
   background-color: $blue;
   color: $white;
   font-size: 1.4rem;
