+++ src/scss/galleryPage.scss
@@ -1,5 +1,5 @@
-@use '../../abstracts' as *;
-@use '../../base' as *;
+@use './abstracts' as *;
+@use './base' as *;
 
 .gallery {
   width: 100%;
