+++ src/scss/landingPage.scss
@@ -3,5 +3,6 @@
 
 .landingContent {
   display: flex;
+  min-height: calc(100vh - 60px);
   background-color: $lightGray;
 }
