+++ src/scss/mainPage.scss
@@ -3,7 +3,7 @@
 #mainPage {
   position: relative;
   width: 100%;
-  height: calc(100vh - 60px);
+  // height: calc(100vh - 60px);
   box-sizing: border-box;
   padding: 1rem 4rem;
   .mainPage {
@@ -17,11 +17,10 @@
       &-wrap {
         display: grid;
         grid-template-columns: repeat(4, 1fr);
-        grid-template-rows: repeat(3, 1fr);
+        grid-template-rows: repeat(2, 1fr);
         row-gap: 1rem;
         column-gap: 1.6rem;
         width: 100%;
-        height: 100%;
       }
       &-box {
         display: flex;
@@ -38,6 +37,10 @@
           &-gallery {
             grid-column: 3 / span 2;
             grid-row: 2 / span 2;
+
+            a {
+              height: 100%;
+            }
           }
         }
       }
